+++ com.unity.ml-agents/Runtime/Sensors/Reflection/ReflectionSensorBase.cs
@@ -91,9 +91,9 @@ namespace Unity.MLAgents.Sensors.Reflection
         public void Reset() { }
 
         /// <inheritdoc/>
-        public SensorCompressionType GetCompressionType()
+        public CompressionSpec GetCompressionSpec()
         {
-            return SensorCompressionType.None;
+            return CompressionSpec.Default();
         }
 
         /// <inheritdoc/>