+++ com.unity.ml-agents/Runtime/Sensors/RenderTextureSensor.cs
@@ -88,9 +88,9 @@ namespace Unity.MLAgents.Sensors
         public void Reset() { }
 
         /// <inheritdoc/>
-        public SensorCompressionType GetCompressionType()
+        public CompressionSpec GetCompressionSpec()
         {
-            return m_CompressionType;
+            return new CompressionSpec(m_CompressionType);
         }
 
         /// <inheritdoc/>